// app/common/constants/asset-type.ts
enum AssetType {
  VIDEO = 'video',
  PHOTO = 'photo',
  TEXT = 'text',
  TEXT_V = 'textV',
  SCROLL_TEXT = 'scrollText',
  SCROLL_TEXT_V = 'scrollTextV',
  WEB = 'web',
  QR = 'qr',
  MESSAGE_BOARD = 'messageboard',
  SLIDESHOW = 'slideshow',

  DISPLAY_DATE = 'display_date',
  DISPLAY_TIME = 'display_time',
  COUNT_DATE = 'count_date',
}

export default AssetType;